+++ axelor-production/src/main/java/com/axelor/apps/production/service/MpsWeeklyScheduleServiceImpl.java
@@ -1,0 +1,53 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2019 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or  modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.axelor.apps.production.service;
+
+import com.axelor.apps.production.db.MpsWeeklySchedule;
+import java.math.BigDecimal;
+
+public class MpsWeeklyScheduleServiceImpl implements MpsWeeklyScheduleService {
+
+  @Override
+  public void countTotalHours(MpsWeeklySchedule mpsWeeklySchedule) {
+    BigDecimal totalHours = BigDecimal.ZERO;
+
+    if (mpsWeeklySchedule.getIsMonday()) {
+      totalHours = totalHours.add(mpsWeeklySchedule.getHoursMonday());
+    }
+    if (mpsWeeklySchedule.getIsTuesday()) {
+      totalHours = totalHours.add(mpsWeeklySchedule.getHoursTuesday());
+    }
+    if (mpsWeeklySchedule.getIsWednesday()) {
+      totalHours = totalHours.add(mpsWeeklySchedule.getHoursWednesday());
+    }
+    if (mpsWeeklySchedule.getIsThursday()) {
+      totalHours = totalHours.add(mpsWeeklySchedule.getHoursThursday());
+    }
+    if (mpsWeeklySchedule.getIsFriday()) {
+      totalHours = totalHours.add(mpsWeeklySchedule.getHoursFriday());
+    }
+    if (mpsWeeklySchedule.getIsSaturday()) {
+      totalHours = totalHours.add(mpsWeeklySchedule.getHoursSaturday());
+    }
+    if (mpsWeeklySchedule.getIsSunday()) {
+      totalHours = totalHours.add(mpsWeeklySchedule.getHoursSunday());
+    }
+
+    mpsWeeklySchedule.setTotalHours(totalHours);
+  }
+}